acceptance: do not use 127.0.0.1 for host
reason: We run test/docker on the same host that wazo-platform. The
wazo_host value is used to indicate on which host linphone must
register. Since linphone run on a docker, if we set 127.0.0.1, the
REGISTER will never go out the container ...
This is why we set wazo_host on a valid ipv4 that redirect on the
host